+++ axum/src/routing/mod.rs
@@ -5,21 +5,19 @@ use crate::{
     body::{Body, HttpBody},
     extract::connect_info::IntoMakeServiceWithConnectInfo,
     response::Response,
-    routing::strip_prefix::StripPrefix,
     util::try_downcast,
 };
 use axum_core::response::IntoResponse;
 use http::Request;
 use matchit::MatchError;
 use std::{
-    borrow::Cow,
     collections::HashMap,
     convert::Infallible,
     fmt,
     sync::Arc,
     task::{Context, Poll},
 };
-use tower::{layer::layer_fn, util::MapResponseLayer, ServiceBuilder};
+use tower::{util::MapResponseLayer, ServiceBuilder};
 use tower_layer::Layer;
 use tower_service::Service;
 
@@ -65,7 +63,6 @@ pub struct Router<B = Body> {
     routes: HashMap<RouteId, Endpoint<B>>,
     node: Arc<Node>,
     fallback: Fallback<B>,
-    nested_at_root: bool,
 }
 
 impl<B> Clone for Router<B> {
@@ -74,7 +71,6 @@ impl<B> Clone for Router<B> {
             routes: self.routes.clone(),
             node: Arc::clone(&self.node),
             fallback: self.fallback.clone(),
-            nested_at_root: self.nested_at_root,
         }
     }
 }
@@ -94,7 +90,6 @@ impl<B> fmt::Debug for Router<B> {
             .field("routes", &self.routes)
             .field("node", &self.node)
             .field("fallback", &self.fallback)
-            .field("nested_at_root", &self.nested_at_root)
             .finish()
     }
 }
@@ -115,7 +110,6 @@ where
             routes: Default::default(),
             node: Default::default(),
             fallback: Fallback::Default(Route::new(NotFound)),
-            nested_at_root: false,
         }
     }
 
@@ -177,7 +171,7 @@ where
         let mut node =
             Arc::try_unwrap(Arc::clone(&self.node)).unwrap_or_else(|node| (*node).clone());
         if let Err(err) = node.insert(path, id) {
-            self.panic_on_matchit_error(err);
+            panic!("Invalid route {path:?}: {err}");
         }
         self.node = Arc::new(node);
     }
@@ -200,63 +194,22 @@ where
 
         let prefix = path;
 
-        if path == "/" {
-            self.nested_at_root = true;
-        }
+        let path = if path.ends_with('/') {
+            format!("{path}*{NEST_TAIL_PARAM}")
+        } else {
+            format!("{path}/*{NEST_TAIL_PARAM}")
+        };
 
-        match try_downcast::<Router<B>, _>(svc) {
-            // if the user is nesting a `Router` we can implement nesting
-            // by simplying copying all the routes and adding the prefix in
-            // front
-            Ok(router) => {
-                let Router {
-                    mut routes,
-                    node,
-                    fallback,
-                    // nesting a router that has something nested at root
-                    // doesn't mean something is nested at root in _this_ router
-                    // thus we don't need to propagate that
-                    nested_at_root: _,
-                } = router;
+        let svc = strip_prefix::StripPrefix::new(svc, prefix);
+        self = self.route(&path, svc.clone());
 
-                if let Fallback::Custom(_) = fallback {
-                    panic!("Cannot nest `Router`s that has a fallback");
-                }
-
-                for (id, nested_path) in &node.route_id_to_path {
-                    let route = routes.remove(id).unwrap();
-                    let full_path: Cow<str> = if &**nested_path == "/" {
-                        path.into()
-                    } else if path == "/" {
-                        (&**nested_path).into()
-                    } else if let Some(path) = path.strip_suffix('/') {
-                        format!("{}{}", path, nested_path).into()
-                    } else {
-                        format!("{}{}", path, nested_path).into()
-                    };
-                    self = match route {
-                        Endpoint::MethodRouter(method_router) => self.route(
-                            &full_path,
-                            method_router.layer(layer_fn(|s| StripPrefix::new(s, prefix))),
-                        ),
-                        Endpoint::Route(route) => {
-                            self.route(&full_path, StripPrefix::new(route, prefix))
-                        }
-                    };
-                }
-
-                debug_assert!(routes.is_empty());
-            }
-            // otherwise we add a wildcard route to the service
-            Err(svc) => {
-                let path = if path.ends_with('/') {
-                    format!("{}*{}", path, NEST_TAIL_PARAM)
-                } else {
-                    format!("{}/*{}", path, NEST_TAIL_PARAM)
-                };
-
-                self = self.route(&path, strip_prefix::StripPrefix::new(svc, prefix));
-            }
+        // `/*rest` is not matched by `/` so we need to also register a router at the
+        // prefix itself. Otherwise if you were to nest at `/foo` then `/foo` itself
+        // wouldn't match, which it should
+        self = self.route(prefix, svc.clone());
+        if !prefix.ends_with('/') {
+            // same goes for `/foo/`, that should also match
+            self = self.route(&format!("{prefix}/"), svc);
         }
 
         self
@@ -271,7 +224,6 @@ where
             routes,
             node,
             fallback,
-            nested_at_root,
         } = other.into();
 
         for (id, route) in routes {
@@ -293,8 +245,6 @@ where
                 panic!("Cannot merge two `Router`s that both have a fallback")
             }
         };
-
-        self.nested_at_root = self.nested_at_root || nested_at_root;
 
         self
     }
@@ -334,7 +284,6 @@ where
             routes,
             node: self.node,
             fallback,
-            nested_at_root: self.nested_at_root,
         }
     }
 
@@ -371,7 +320,6 @@ where
             routes,
             node: self.node,
             fallback: self.fallback,
-            nested_at_root: self.nested_at_root,
         }
     }
 
@@ -472,17 +420,6 @@ where
         match &mut route {
             Endpoint::MethodRouter(inner) => inner.call(req),
             Endpoint::Route(inner) => inner.call(req),
-        }
-    }
-
-    fn panic_on_matchit_error(&self, err: matchit::InsertError) {
-        if self.nested_at_root {
-            panic!(
-                "Invalid route: {}. Note that `nest(\"/\", _)` conflicts with all routes. Use `Router::fallback` instead",
-                err,
-            );
-        } else {
-            panic!("Invalid route: {}", err);
         }
     }
 }